# docs/swagger.json
{
  "consumes": [
    "application/vnd.api+json",
    "application/json"
  ],
  "produces": [
    "application/vnd.api+json",
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "the purpose of this service is to provide a template for Go-based\nmicroservice development. This can be used to start development on\na new service or as an example of Optii accepted best practices for\ndeveloping Go-based microservices",
    "title": "Template Service for Go-based Development",
    "termsOfService": "there are no TOS at this moment, use at your own risk we take no responsibility",
    "version": "0.0.1"
  },
  "host": "localhost",
  "basePath": "/api",
  "paths": {
...
    "/template": {
      "get": {
        "tags": [
          "entries"
        ],
        "operationId": "listTemplates"
      },
      "post": {
        "description": "returns new Template",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "entries"
        ],
        "summary": "Create new Template",
        "operationId": "createTemplate",
        "parameters": [
          {
            "description": "Template object that needs to be added",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/GlobalTemplateStruct"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/GlobalTemplateStruct"
            }
          }
        }
      }
    },
    "/template/{id}": {
      "get": {
        "description": "returns Template",
        "produces": [
          "application/json"
        ],
        "tags": [
          "entries"
        ],
        "summary": "Get Template by given id",
        "operationId": "getTemplate",
        "parameters": [
          {
            "type": "number",
            "description": "Numeric ID of the Template to get",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "$ref": "#/definitions/GlobalTemplateStruct"
            }
          }
        }
      },
      "put": {
        "tags": [
          "entries"
        ],
        "operationId": "updateTemplate"
      },
      "delete": {
        "description": "Numeric ID of the Template to delete",
        "tags": [
          "entries"
        ],
        "operationId": "deleteTemplate"
      }
    }
  },
...
  "securityDefinitions": {
    "optii_apikey": {
      "type": "apiKey",
      "name": "KEY",
      "in": "header"
    },
    "optii_oauth2": {
      "description": "example",
      "type": "oauth2",
      "flow": "accessCode",
      "authorizationUrl": "https://localhost/oauth2/auth",
      "tokenUrl": "https://localhost/oauth2/token",
      "scopes": {
        "read_scope": "description here",
        "write_scope": "description here"
      }
    }
  },
  "security": [
    {
      "optii_apikey": []
    },
    {
      "optii_oauth2": []
    }
  ]
}